Made confirmation code entry page easier to navigate to for user
In index.html, added a hyperlink that goes right to the Confirmation Code Entry Page.

In email.php, added a hyperlink in the email that goes out to the entered email address. From the email, the user can click on the link and go directly to the Confirmation Code Entry Page