Added Functionality for Choosing Class in EvalForm and passed real values instead of dummies
Choosing the class is a dropdown menu that refreshes the page and loads the students from that class into the teammate radio buttons for evaluation submission. I had previously used a bunch of dummy values for testing the form submission but now all are passed in except the team name.